[CI] update CI docker to ubuntu20.04
- Simplify the CI pipeline by removing some unnecessary testing items.
- Fix the hiredis dependency to tag v1.1.0

+++ tensorflow_recommenders_addons/dynamic_embedding/python/kernel_tests/dynamic_embedding_ops_test.py
@@ -66,7 +66,7 @@ def _type_converter(tf_type):
   mapper = {
       dtypes.int32: np.int32,
       dtypes.int64: np.int64,
-      dtypes.float32: np.float,
+      dtypes.float32: float,
       dtypes.float64: np.float64,
   }
   return mapper[tf_type]
@@ -383,9 +383,8 @@ class EmbeddingLookupTest(test.TestCase):
                 initializer=2.0,
                 dim=dim,
             )
-            ids = np.random.randint(2**31,
-                                    size=np.prod(ids_shape),
-                                    dtype=np.int).reshape(ids_shape)
+            ids = np.random.randint(2**31, size=np.prod(ids_shape),
+                                    dtype=int).reshape(ids_shape)
             ids = constant_op.constant(ids, dtype=dtypes.int64)
             simple = params.lookup(ids)
             self.evaluate(params.upsert(ids, simple))
@@ -414,9 +413,8 @@ class EmbeddingLookupTest(test.TestCase):
             params_nn = variable_scope.get_variable("n",
                                                     shape=[100, dim],
                                                     use_resource=False)
-            ids = np.random.randint(2**31,
-                                    size=np.prod(ids_shape),
-                                    dtype=np.int).reshape(ids_shape)
+            ids = np.random.randint(2**31, size=np.prod(ids_shape),
+                                    dtype=int).reshape(ids_shape)
             ids = constant_op.constant(ids, dtype=dtypes.int64)
 
             embedding_test = de.embedding_lookup(params, ids)
